modified:   TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Rover/HardwareDummyRover.java
	modified:   TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Rover/RoverAuto.java

Matthew worked on seeing the target while the robot was turning and stopping.
It worked as long as robot did not turn too fast and wasn't too far away from the topic
Aligned the camera position with the robot.
Started working to incorporate IMU program to go straight from the heading received.